Taming EIEIO a bit more
